+++ scripts/main.js
@@ -271,6 +271,7 @@ app.on('ready', () => {
       newVer: { type: 'boolean', alias: 'V', desc: 'Check for new release', default: undefined }, // has a setting
       select: { type: 'string', alias: 's', desc: 'Selection criteria', default: undefined },
       idOnly: { type: 'boolean', alias: 'i', desc: 'Search id only', default: false },
+      regex: { type: 'boolean', alias: 'x', desc: 'Regex Search', default: false },
       vql: { type: 'boolean', alias: 'q', desc: 'VQL search', default: false },
       limitDepth: { type: 'boolean', alias: '1', desc: 'Limit reachable nodes to 1 level', default: false }, //rq: ->(rq_limited_walk_cl)
       exclIds: { type: 'string', alias: 'e', desc: 'Excluded ids, comma separated', default: undefined },